test: Move new code to Footer.text.tsx
